+++ src/Message/Authentication/AccountNotActiveMessage.php
@@ -1,0 +1,25 @@
+<?php
+/**
+ * @author    Andrew Coulton <andrew@ingenerator.com>
+ * @licence   proprietary
+ */
+
+namespace Ingenerator\Warden\UI\Kohana\Message\Authentication;
+
+
+use Ingenerator\Pigeonhole\Message;
+use Ingenerator\Pigeonhole\Message\KohanaMessage;
+
+class AccountNotActiveMessage extends KohanaMessage
+{
+    public function __construct($email)
+    {
+        parent::__construct(
+            'warden_flash_messages',
+            'authentication.account_not_active',
+            ['%email%' => $email],
+            Message::DANGER
+        );
+    }
+
+}