+++ src/Message/Register/AccountActivatedMessage.php
@@ -1,0 +1,24 @@
+<?php
+/**
+ * @author    Andrew Coulton <andrew@ingenerator.com>
+ * @licence   proprietary
+ */
+
+namespace Ingenerator\Warden\UI\Kohana\Message\Register;
+
+
+use Ingenerator\Pigeonhole\Message;
+use Ingenerator\Pigeonhole\Message\KohanaMessage;
+
+class AccountActivatedMessage extends KohanaMessage
+{
+    public function __construct()
+    {
+        parent::__construct(
+            'warden_flash_messages',
+            'register.account_activated',
+            [],
+            Message::SUCCESS
+        );
+    }
+}