+++ src/Message/Register/InvalidActivationLinkMessage.php
@@ -1,0 +1,24 @@
+<?php
+/**
+ * @author    Andrew Coulton <andrew@ingenerator.com>
+ * @licence   proprietary
+ */
+
+namespace Ingenerator\Warden\UI\Kohana\Message\Register;
+
+
+use Ingenerator\Pigeonhole\Message;
+
+class InvalidActivationLinkMessage extends Message\KohanaMessage
+{
+    public function __construct()
+    {
+        parent::__construct(
+            'warden_flash_messages',
+            'register.invalid_activation_link',
+            [],
+            Message::DANGER
+        );
+    }
+
+}